Implemented Group system
Loosely based on the entt ECS system. In theory much better performance as groups should reduce cache misses. Unfortunately I think I've kinda butchered the concept. Some intricacies of templates and functors are beyond me at the moment.  Sending templated functions as signals to be invoked every time a component that is part of a group is created or destroyed is simply not something I understand, and I don't want to just copy "entt" source code